+++ docs/changelog.rst
@@ -14,8 +14,10 @@ Added
 -----
 - config component ``WflowConfigComponent``: represents the Wflow configuration TOML file.
 - staticmaps component ``WflowStaticMapsComponent``: represents Wflow static and cyclic data (used to be grid).
+- forcing component ``WflowForcingComponent``: represents Wflow input forcing data.
 - states component ``WflowStatesComponent``: represents Wflow input states data.
 - geoms component ``WflowGeomsComponent``: represents Wflow staticgeoms data.
+- output_grid ``WflowOutputGridComponent``, output_scalar ``WflowOutputScalarComponent`` and output_csv ``WflowOutputCSVComponent`` components: reprensents Wflow outputs (used to be results).
 - **write_geoms**: added function arguments ``to_wgs84``  to convert the geometry to WGS84 before writing it to file. PR #432
 
 Changed
@@ -27,6 +29,7 @@ Changed
   Eg wflow.config.data, wflow.staticmaps.data
 - The names of some of the model components have changed: ``grid`` to ``staticmaps``.
 - **setup_config**: the method now explicitly uses a dictionary with the options to add/update.
+- **utils.read_csv_results** has been renamed to **utils.read_csv_output**.
 - Reverted the use of TOMLkit dependency